@@ -1,5 +1,8 @@
 {
   "name": "@igo/demo",
+  "workspaces": [
+    "./packages/*"
+  ],
   "version": "1.14.2",
   "description": "IGO Library",
   "author": "IGO Community",
@@ -76,16 +79,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^14.1.2",
-    "@angular/cdk": "^14.1.2",
-    "@angular/common": "^14.1.2",
-    "@angular/compiler": "^14.1.2",
-    "@angular/core": "^14.1.2",
-    "@angular/forms": "^14.1.2",
-    "@angular/material": "^14.1.2",
-    "@angular/platform-browser": "^14.1.2",
-    "@angular/platform-browser-dynamic": "^14.1.2",
-    "@angular/router": "^14.1.2",
+    "@angular/animations": "^14.3.0",
+    "@angular/cdk": "^14.2.7",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/material": "^14.2.7",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/router": "^14.3.0",
     "@azure/msal-angular": "^2.4.1",
     "@azure/msal-browser": "^2.28.1",
     "@mat-datetimepicker/core": "~10.0.0",
@@ -95,7 +97,7 @@
     "@turf/helpers": "^6.5.0",
     "@turf/line-intersect": "^6.5.0",
     "@turf/point-on-feature": "^6.5.0",
-    "ajv": "^8.11.0",
+    "ajv": "^6.9.1",
     "angular-shepherd": "14.0.0",
     "bowser": "^2.10.0",
     "core-js": "^3.24.1",
@@ -127,14 +129,15 @@
     "zone.js": "~0.11.7"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^14.1.2",
+    "@angular-devkit/build-angular": "^14.2.11",
     "@angular-eslint/builder": "14.0.2",
     "@angular-eslint/eslint-plugin": "14.0.2",
     "@angular-eslint/eslint-plugin-template": "14.0.2",
     "@angular-eslint/schematics": "14.0.2",
     "@angular-eslint/template-parser": "14.0.2",
-    "@angular/cli": "^14.1.2",
-    "@angular/compiler-cli": "^14.1.2",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/cli": "^14.2.10",
+    "@angular/compiler-cli": "^14.3.0",
     "@babel/core": "^7.18.10",
     "@babel/preset-env": "^7.18.10",
     "@compodoc/compodoc": "^1.1.15",
@@ -171,7 +174,7 @@
     "karma-coverage": "^2.2.0",
     "karma-jasmine": "~4.0.1",
     "karma-jasmine-html-reporter": "^1.7.0",
-    "ng-packagr": "^14.1.0",
+    "ng-packagr": "^14.2.2",
     "npm-run-all": "^4.1.5",
     "protractor": "~7.0.0",
     "sass": "^1.54.4",
@@ -180,5 +183,8 @@
     "wait-on": "^6.0.0",
     "webdriver-manager": "^12.1.6",
     "webpack": "^5.74.0"
+  },
+  "engines": {
+    "node": ">=16.19.0"
   }
 }
