@@ -20,15 +20,19 @@
     "tslib": "^2.4.0"
   },
   "peerDependencies": {
-    "@angular/cdk": "^14.1.2",
-    "@angular/common": "^14.1.2",
-    "@angular/core": "^14.1.2",
-    "@angular/forms": "^14.1.2",
-    "@angular/material": "^14.1.2",
+    "@angular/cdk": "^14.2.7",
+    "@angular/common": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/material": "^14.2.7",
+    "@angular/router": "^14.3.0",
     "@igo2/core": "^1.14.2",
     "@igo2/utils": "^1.14.2",
     "rxjs": "^7.5.6",
     "@azure/msal-angular": "^2.4.1",
     "@azure/msal-browser": "^2.28.1"
+  },
+  "engines": {
+    "node": ">=16.19.0"
   }
 }
