@@ -14,9 +14,10 @@
     "url": "https://github.com/infra-geo-ouverte/igo2-lib/issues"
   },
   "peerDependencies": {
-    "@angular/common": "^14.1.2",
-    "@angular/core": "^14.1.2",
-    "@angular/material": "^14.1.2",
+    "@angular/common": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/material": "^14.2.7",
+    "@angular/platform-browser": "^14.3.0",
     "@igo2/core": "^1.14.2",
     "@igo2/utils": "^1.14.2",
     "scroll-into-view-if-needed": "^2.2.20",
@@ -24,5 +25,8 @@
   },
   "dependencies": {
     "tslib": "^2.4.0"
+  },
+  "engines": {
+    "node": ">=16.19.0"
   }
 }
