@@ -24,11 +24,12 @@
     "windows-1252": "^3.0.4"
   },
   "peerDependencies": {
-    "@angular/animations": "^14.1.2",
-    "@angular/common": "^14.1.2",
-    "@angular/core": "^14.1.2",
-    "@angular/forms": "^14.1.2",
-    "@angular/material": "^14.1.2",
+    "@angular/animations": "^14.3.0",
+    "@angular/common": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/material": "^14.2.7",
+    "@angular/platform-browser": "^14.3.0",
     "@igo2/common": "^1.14.2",
     "@igo2/core": "^1.14.2",
     "@igo2/utils": "^1.14.2",
@@ -37,7 +38,8 @@
     "proj4": "^2.8.0",
     "rxjs": "^7.5.6",
     "striptags": "^3.2.0",
-    "ngx-color-picker": "^13.0.0"
+    "ngx-color-picker": "^13.0.0",
+    "ngx-indexed-db": "^11.0.2"
   },
   "optionalDependencies": {
     "file-saver": "^2.0.2",
@@ -46,5 +48,8 @@
     "jszip": "^3.10.1",
     "moment": "^2.29.4",
     "stream": "^0.0.2"
+  },
+  "engines": {
+    "node": ">=16.19.0"
   }
 }
