@@ -1,7 +1,7 @@
 import { Injectable, Optional } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Observable, of } from 'rxjs';
-import { map, catchError } from 'rxjs/operators';
+import { combineLatest, Observable, of } from 'rxjs';
+import { map, catchError, concatMap } from 'rxjs/operators';
 import {stylefunction} from "ol-mapbox-style";
 import { AuthInterceptor } from '@igo2/auth';
 import { ObjectUtils } from '@igo2/utils';
@@ -39,7 +39,8 @@ import {
   VectorLayerOptions,
   AnyLayerOptions,
   VectorTileLayer,
-  VectorTileLayerOptions
+  VectorTileLayerOptions,
+  LayerOptions
 } from './layers';
 
 import { computeMVTOptionsOnHover } from '../utils/layer.utils';
@@ -47,6 +48,7 @@ import { StyleService } from '../../style/style-service/style.service';
 import { MessageService } from '@igo2/core';
 import { GeoNetworkService } from '../../offline/shared/geo-network.service';
 import { StyleLike as OlStyleLike } from 'ol/style/Style';
+import { LayerDBService } from '../../offline/layerDB/layerDB.service';
 
 @Injectable({
   providedIn: 'root'
@@ -56,8 +58,9 @@ export class LayerService {
     private http: HttpClient,
     private styleService: StyleService,
     private dataSourceService: DataSourceService,
-    private geoNetwork: GeoNetworkService,
+    private geoNetworkService: GeoNetworkService,
     private messageService: MessageService,
+    private layerDBService: LayerDBService,
     @Optional() private authInterceptor: AuthInterceptor
   ) {}
 
@@ -146,9 +149,11 @@ export class LayerService {
     const legacyStyleOptions = ['styleByAttribute', 'hoverStyle', 'mapboxStyle', 'clusterBaseStyle', 'style'];
     // handling legacy property.
     this.handleLegacyStyles(layerOptions, legacyStyleOptions);
-
-    if (layerOptions.igoStyle.igoStyleObject) {
+    if (layerOptions.igoStyle.igoStyleObject && !layerOptions.idbInfo?.storeToIdb) {
       style = (feature, resolution) => this.styleService.createStyle(layerOptions.igoStyle.igoStyleObject, feature, resolution);
+    } else if (layerOptions.igoStyle.igoStyleObject && layerOptions.idbInfo?.storeToIdb) {
+      // temporary fix todo : handle it with geostyler.
+      style = this.styleService.parseStyle('style', layerOptions.igoStyle.igoStyleObject);
     }
 
     if (layerOptions.source instanceof ArcGISRestDataSource) {
@@ -163,7 +168,13 @@ export class LayerService {
           resolution
         );
       };
-      igoLayer = new VectorLayer(layerOptions, this.messageService, this.authInterceptor, this.geoNetwork);
+      igoLayer = new VectorLayer(
+        layerOptions,
+        this.messageService,
+        this.authInterceptor,
+        this.geoNetworkService,
+        this.geoNetworkService.geoDBService,
+        this.layerDBService);
     }
 
     if (layerOptions.source instanceof ClusterDataSource) {
@@ -177,7 +188,13 @@ export class LayerService {
           baseStyle
         );
       };
-      igoLayer = new VectorLayer(layerOptions, this.messageService, this.authInterceptor, this.geoNetwork);
+      igoLayer = new VectorLayer(
+        layerOptions,
+        this.messageService,
+        this.authInterceptor,
+        this.geoNetworkService,
+        this.geoNetworkService.geoDBService,
+        this.layerDBService);
     }
 
     const layerOptionsOl = Object.assign({}, layerOptions, {
@@ -185,7 +202,12 @@ export class LayerService {
     });
 
     if (!igoLayer) {
-      igoLayer = new VectorLayer(layerOptionsOl, this.messageService, this.authInterceptor, this.geoNetwork);
+      igoLayer = new VectorLayer(layerOptionsOl,
+        this.messageService,
+        this.authInterceptor,
+        this.geoNetworkService,
+        this.geoNetworkService.geoDBService,
+        this.layerDBService);
     }
 
     this.applyMapboxStyle(igoLayer, layerOptionsOl as any);
@@ -297,6 +319,17 @@ export class LayerService {
         return source + "/" + url;
       }
     }
+  }
+
+  createAsyncIdbLayers(contextUri: string = '*'): Observable<Layer[]> {
+    return this.layerDBService.getAll().pipe(
+      concatMap(res => {
+        const idbLayers = contextUri !== '*' ? res.filter(l => l.detailedContextUri === contextUri) : res;
+        const layersOptions: LayerOptions[] =
+          idbLayers.map(idbl => Object.assign({}, idbl.layerOptions, { sourceOptions: idbl.sourceOptions }));
+        return combineLatest(layersOptions.map(layerOptions => this.createAsyncLayer(layerOptions)));
+      })
+    );
   }
 
 }
